PROJECT DESCRIPTION
Tea Cozy - Fictional Tea Shop Website
This project is a responsive website for a fictional tea shop called "Tea Cozy." It was created based on a design spec, which provided detailed instructions on the layout and styling. The site showcases different teas and features sections such as the company’s mission, locations, and a featured tea collection.

Technologies Used:

HTML5
CSS3
Project Highlights:

Implemented a pixel-perfect design based on a provided spec.
Created a responsive layout using Flexbox.
Used CSS for styling and positioning elements according to the design spec.